:bug: Fix uncaught invariant violation console errors
When loading the Events tab as an admin, uncaught invariant violation console errors are emitte
d. This is due to the `Cell` delegate of the  `Start` column definition for ReactTable. The Cell delegate is required to re
turn either a null or react element. However it was just returning a string. The fix was to wrap the string inside a span.