:bug: Don't set defaultPageSize when showPagination=false
The ReactTable component was set with showPagination={false} and defaultPageSize={events.length}. The component seems to treat these two settings independently and so this combination caused the table to restrict the events shown to the first event list size, but not provide a way to view the next page.

This is a problem if the first list has a small number of events e.g. 0 or 1